refactor(home): more reasonable layout (close: #31)
