+++ src/components/Device.vue
